refactor: major refactoring of the code base
`opts` was renamed to `cli` to conform to upstreams (clap) naming
convention

`lib` was renamed to `utils`

The whole style module was split up - the styles as defined in the theme
are now part of the theme module. The `utils` module contains a simple
method to convert the style definition from the theme to a string.

`theme` is now a submodule of `config`

The `DateProperty` is now part of the theme as well.

The `date`, `calendar` and `agenda` modules where moved to a new
`output` module.