fix: implement parsing of ical dates with timezones
Up until now, carl ignored timezone settings in date properties of ical
files. This commit adds a check if the datetimstring ends with a `Z` and
then parses it as UTC datetime; carl now also checks if there is a TZID
parameter added to the datetimestring and if so, parses the
datetimestring using this timezone.
This adds chrono-tz as dependency for timezone parsing.