feat(cli)!: allow exclusive actions
This is a backwards incompatible change, because actions are now
exclusive. Up until now, `-a` showed the calendar *and* the agenda. The
calendar is still the default, but if you choose to display the agenda
with `-a` it displays the agenda *instead* of the calendar. If you want
both, you have to use `-ac` or `--calendar --agenda`.

Closes: #97